extended message read

+++ src/com/cosylab/epics/caj/impl/requests/SubscriptionUpdateRequest.java
@@ -35,7 +35,7 @@ public class SubscriptionUpdateRequest extends AbstractCARequest {
 		super(transport);
 
 		requestMessage = insertCAHeader(transport, null,
-										(short)15, (short)0, (short)dataType, (short)dataCount,
+										(short)15, 0, (short)dataType, dataCount,
 										sid, ioid);
 	}
